+++ src/main/java/org.mirrors/PathTracer.java
@@ -23,7 +23,7 @@ public class PathTracer extends Renderer{
     public PathTracer(World world) {
         super(world);
         this.pcg = new PCG();
-        this.numOfRays = 5;
+        this.numOfRays = 10;
         this.maxDepth = 2;
         this.russianRouletteLimit = 3;
     }